requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
boolean.py==5.0
boto3==1.38.24
botocore==1.38.24
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2
django-appconf==1.1.0
django-axes==8.0.0
django-background-tasks==1.2.8
django-import-export==4.3.7
django-storages==1.14.6
djangorestframework==3.16.0
dj-database-url==2.3.0
filelock==3.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jmespath==1.0.1
license-expression==30.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
oauthlib==3.2.2
packageurl-python==0.16.0
packaging==25.0
pilkit==3.0
pillow==11.1.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
py-serializable==2.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
stripe==12.1.0
tablib==3.8.0
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.1.0
websockets==15.0.1
whitenoise==6.9.0
channels==4.1.0
daphne==4.1.2

channels-redis==4.2.0

# Security headers
django-csp==3.8
